Automatic cloud detection and elements retrieval
Targeting EC2 support for now, but supporting other cloud metadata
services should be easy to implement.

Use the cloudsysfs library to quickly determine if we're on a cloud
provider.

If we're an EC2 instance, assume the http://169.254.169.254/latest
endpoint is available (also a pretty reliable bet). Crawl this
recursively for meta and user data, populating the "ec2" top-level
element path.